🔧 Moved supported locales to i18n package with Typescript support
