+++ apps/admin-x-settings/package.json
@@ -36,6 +36,7 @@
   "dependencies": {
     "@codemirror/lang-html": "6.4.9",
     "@tryghost/color-utils": "0.2.10",
+    "@tryghost/i18n": "0.0.0",
     "@tryghost/kg-unsplash-selector": "0.3.6",
     "@tryghost/limit-service": "1.4.1",
     "@tryghost/nql": "0.12.7",
